[5주차] 백준_골드5_15486
